creating seperate components for footer items
