+++ src/components/footer/FooterCopyRight.js
@@ -1,0 +1,18 @@
+import React from 'react'
+
+
+const CRFooter = () => 
+(
+    <div className="copyright">
+    <div className="container">
+      <p className="small"><a href="https://my.sparkol.com" target="_blank" rel="noopener noreferrer">Sparkol Account</a></p>
+      <ul>
+        <li className="small">© 2018 Sparkol Limited, registered in England and Wales, no 06762963</li>
+        <li className="small">1 Temple Way 5th Floor, Bristol, BS2 0BY, UK</li>
+      </ul>
+      <p className="small"><a href="http://www.sparkol.com/legal/" target="_blank" rel="noopener noreferrer">Privacy and cookies</a></p>
+    </div>
+  </div>
+)
+
+export default CRFooter
