+++ src/components/footer/FooterSocial.js
@@ -1,0 +1,20 @@
+import React from 'react'
+
+
+
+const SFooter = () => 
+(
+    <div className="social">
+    <div className="container">
+      <ul className="menu">
+        <li><a href="https://twitter.com/SparkolHQ" className="icon-social icon-twitter" target="_blank" rel="noopener noreferrer">Twitter</a></li>
+        <li><a href="https://www.facebook.com/SparkolHQ" className="icon-social icon-facebook" target="_blank" rel="noopener noreferrer">Facebook</a></li>
+        <li><a href="https://www.youtube.com/channel/UC-M3N1dBICkLMU8fTlmpgjw" className="icon-social icon-youtube" target="_blank" rel="noopener noreferrer">YouTube</a></li>
+        <li><a href="https://plus.google.com/u/0/+Sparkol/posts" className="icon-social icon-gplus" target="_blank" rel="noopener noreferrer">Google+</a></li>
+        <li><a href="https://www.linkedin.com/company/2755634" className="icon-social icon-linkedin" target="_blank" rel="noopener noreferrer">LinkedIn</a></li>
+        <li><a href="https://instagram.com/sparkolhq/" className="icon-social icon-instagram" target="_blank" rel="noopener noreferrer">Instagram</a></li>
+      </ul>
+    </div>
+  </div>
+)
+export default SFooter
